feat(User) have ctreated a new package 'ui.user', moved files from 'main' to 'ui.user' which connected with 'user'
