+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:supportsRtl="true"
         android:hardwareAccelerated="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".main.MainActivity">
+        <activity android:name=".ui.user.UserActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -27,8 +27,8 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".main.AddUserActivity"
-            android:parentActivityName=".main.MainActivity" />
+            android:name=".ui.user.AddUserActivity"
+            android:parentActivityName=".ui.user.UserActivity" />
         <activity android:name=".request.NewRequestActivity" />
         <activity android:name=".request.ShowRequestsActivity" />
         <activity
