+++ app/src/main/java/com/example/w6dagger/di/module/ActivitiesModule.kt
@@ -1,21 +1,21 @@
 package com.example.w6dagger.di.module
 
-import com.example.w6dagger.gallery.GalleryActivity
 import com.example.w6dagger.camera.CameraActivity
+import com.example.w6dagger.gallery.GalleryActivity
 import com.example.w6dagger.gallery.internalGallery.InternalGalleryActivity
 import com.example.w6dagger.images.FavoriteImagesActivity
 import com.example.w6dagger.images.ImagesOfRequestActivity
-import com.example.w6dagger.main.MainActivity
 import com.example.w6dagger.map.MapActivity
 import com.example.w6dagger.request.NewRequestActivity
 import com.example.w6dagger.request.ShowRequestsActivity
+import com.example.w6dagger.ui.user.UserActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
 @Module
 abstract class ActivitiesModule {
     @ContributesAndroidInjector
-    abstract fun contributeMainActivity(): MainActivity
+    abstract fun contributeMainActivity(): UserActivity
 
     @ContributesAndroidInjector
     abstract fun contributeShowRequestsActivity(): ShowRequestsActivity
