+++ app/src/main/java/com/example/w6dagger/di/module/ViewModelModule.kt
@@ -2,14 +2,14 @@ package com.example.w6dagger.di.module
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.w6dagger.gallery.GalleryActivityViewModel
 import com.example.w6dagger.camera.EditPictureActivityModel
+import com.example.w6dagger.gallery.GalleryActivityViewModel
 import com.example.w6dagger.gallery.internalGallery.InternalGalleryActivityModel
 import com.example.w6dagger.images.ImageViewModel
-import com.example.w6dagger.main.UserViewModel
 import com.example.w6dagger.main.ViewModelKey
 import com.example.w6dagger.main.ViewModelProviderFactory
 import com.example.w6dagger.request.RequestViewModel
+import com.example.w6dagger.ui.user.UserViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
