+++ app/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.example.w6dagger.main.MainActivity">
+        tools:context="com.example.w6dagger.domain.usecase.ui.user.MainActivity">
 
 
         <include
